Make DataWriter, remove old config file
