--- sodym/tools/visualize/__init__.py
@@ -1,3 +1,0 @@
-# make these functions available as attributes of the visualize module
-from .sankey import visualize_mfa_sankey as visualize_mfa_sankey
-from .plot_array import visualize_array as visualize_array
